@@ -4,7 +4,6 @@ import android.content.Intent;
 import android.os.Bundle;
 
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.fragment.app.Fragment;
 
 import android.view.LayoutInflater;
@@ -16,89 +15,101 @@ import android.widget.Toast;
 
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.Firebase;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 
+// This fragment handles user login using Firebase Authentication
 public class LoginFragment extends Fragment {
-    FirebaseAuth auth;
 
+    FirebaseAuth auth; // Firebase authentication object
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
 
-        View view =  inflater.inflate(R.layout.fragment_login, container, false);
+        // Inflate the login fragment layout
+        View view = inflater.inflate(R.layout.fragment_login, container, false);
+
+        // Initialize Firebase authentication
         auth = FirebaseAuth.getInstance();
+
+        // Get references to the email and password input fields and login button
         EditText emailLoginEt = view.findViewById(R.id.editTextEmailLogin);
         EditText passwordLogInEt = view.findViewById(R.id.editTextPasswordLogin);
         Button loginBtn = view.findViewById(R.id.buttonLogin);
+
+        // Set the login button click listener
         loginBtn.setOnClickListener(new View.OnClickListener() {
             @Override
-            public void onClick(View v)
-            {
+            public void onClick(View v) {
+
                 String email = emailLoginEt.getText().toString();
                 String password = passwordLogInEt.getText().toString();
-                if(email.isEmpty() || password.isEmpty())
+
+                // Check if fields are empty
+                if (email.isEmpty() || password.isEmpty()) {
                     Toast.makeText(getContext(), "email and/or password can't be empty", Toast.LENGTH_LONG).show();
+                }
+
+                // Validate email and password fields
                 else if (!validatePassword() || !validateUsername()) {
                     Toast.makeText(getContext(), "username and/or password not valid", Toast.LENGTH_LONG).show();
+                }
 
-                }
-                else{
-                    checkUser();}}
-                public boolean validateUsername(){//יצירת פונקציה הבודקת האם הערך שהכניס ל username לא ריק
-                    String val = emailLoginEt.getText().toString();//מכניס למשתנה את המידע שהכניס המשתמש במסך
-                    if(val.isEmpty()){//תנאי לבדוק אם הכניס מידע ריק
-                        emailLoginEt.setError("email cannot be empty");//מוציא הודעת שגיאה למסך ואומר לו שהערך לא יכול ליהיות ריק
-                        return false;
-                }
-                else{
-                        emailLoginEt.setError(null);//אין שגיאה
-                        return true;
+                // If valid, try to log in
+                else {
+                    checkUser();
                 }
             }
-            public Boolean validatePassword(){//יצירת פונקציה הבודקת האם הערך שהכניס לסיסמה לא ריק
-                String val = passwordLogInEt.getText().toString();//מכניס למשתנה את המידע שהכניס המשתמש במסך
-                if(val.isEmpty()){//תנאי לבדוק אם הכניס מידע ריק
-                    passwordLogInEt.setError("password cannot be empty");//מוציא הודעת שגיאה למסך ואומר לו שהערך לא יכול ליהיות ריק
+
+            // Validate that the email field is not empty
+            public boolean validateUsername() {
+                String val = emailLoginEt.getText().toString();
+                if (val.isEmpty()) {
+                    emailLoginEt.setError("email cannot be empty");
                     return false;
-                }
-                else{
-                    passwordLogInEt.setError(null);//אין שגיאה
+                } else {
+                    emailLoginEt.setError(null); // No error
                     return true;
                 }
             }
-            public void checkUser(){
-                String userUsername = emailLoginEt.getText().toString().trim();//מכניס למשתנה את המידע שהכניס המשתמש במסך
-                String userPassword = passwordLogInEt.getText().toString().trim();//מכניס למשתנה את המידע שהכניס המשתמש במסך
-                auth.signInWithEmailAndPassword(userUsername, userPassword).addOnSuccessListener(new OnSuccessListener<AuthResult>() {//הפעולה בעזרת הפונקציה authentication של firebase שולפת את הנתונים של המשמשים בdatabase ובודק האם הנתונים שהכניס המשתמש למסך נמצאים בdatabase. בנוסף הוא שם מאזים להאם הפעולה תצליח או לא תצליח
-                    @Override
-                    public void onSuccess(AuthResult authResult) {//מה יקרה כאשר הפעולה תצליח והנתונים שהכניס המשתמש כן נמצאו בdatabase
-                        Toast.makeText(getContext(), "good", Toast.LENGTH_SHORT).show();//מדפיס הודעה למסך שעבד
-                        Intent intent = new Intent(getContext(),BuyOrSellActivity.class);//פונקציה המעבירה את המשתמש לעמוד main
-                        startActivity(intent);//קיראה לפונקציה
-                    }
-                }).addOnFailureListener(new OnFailureListener() {//מאזין למקרה שלא נמצאו הנתונים שהכניס המשתמש בdatabase
-                    @Override
-                    public void onFailure(@NonNull Exception e) {//מה יקרה כאשר הפעולה לא תצליח והנתונים שהכניס המשתמש לא נמצאו בdatabase
-                        Toast.makeText(getContext(), "not-good", Toast.LENGTH_SHORT).show();//מדפיס הודעת שגיאה למסך של המשתמש
-                    }
-                });
-                //else
-              //  {
-                  //  String message = FirebaseUtil.LoginToDb(email, password);
-                  //  Toast.makeText(getContext(), message, Toast.LENGTH_LONG).show();
-                 //   if(message.equals("login successfully"))
-                   // {
-                      //  Intent intent = new Intent(getActivity(),BuyOrSellActivity.class);
-                     //   startActivity(intent);
-                   // }
-             //   }
 
+            // Validate that the password field is not empty
+            public Boolean validatePassword() {
+                String val = passwordLogInEt.getText().toString();
+                if (val.isEmpty()) {
+                    passwordLogInEt.setError("password cannot be empty");
+                    return false;
+                } else {
+                    passwordLogInEt.setError(null); // No error
+                    return true;
+                }
+            }
+
+            // Attempt to sign in the user using Firebase Authentication
+            public void checkUser() {
+                String userUsername = emailLoginEt.getText().toString().trim();
+                String userPassword = passwordLogInEt.getText().toString().trim();
+
+                auth.signInWithEmailAndPassword(userUsername, userPassword)
+                        .addOnSuccessListener(new OnSuccessListener<AuthResult>() {
+                            @Override
+                            public void onSuccess(AuthResult authResult) {
+                                // If login is successful, navigate to BuyOrSellActivity
+                                Intent intent = new Intent(getContext(), BuyOrSellActivity.class);
+                                startActivity(intent);
+                            }
+                        })
+                        .addOnFailureListener(new OnFailureListener() {
+                            @Override
+                            public void onFailure(@NonNull Exception e) {
+                                // If login fails, show an error message
+                                Toast.makeText(getContext(), "not-good", Toast.LENGTH_SHORT).show();
+                            }
+                        });
             }
         });
 
-        return view;
+        return view; // Return the view for this fragment
     }
 }
