@@ -9,39 +9,41 @@ import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
 import android.widget.ImageView;
 
-import androidx.activity.EdgeToEdge;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.core.graphics.Insets;
-import androidx.core.view.ViewCompat;
-import androidx.core.view.WindowInsetsCompat;
 
+// This activity serves as a splash screen shown when the app is first launched
 public class MainActivity extends AppCompatActivity {
 
-    ImageView logoIv;
-    Context context;
+    ImageView logoIv;   // Reference to the logo image view
+    Context context;    // Context for the activity
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
 
+        // Set the layout of the splash screen
         setContentView(R.layout.activity_main);
         context = this;
+
+        // Connect the ImageView with the logo in the layout
         logoIv = findViewById(R.id.imageViewLogo);
 
+        // Load the animation from resources and apply it to the logo
         Animation animation = AnimationUtils.loadAnimation(context, R.anim.my_anim);
-        animation.setDuration(5000);
-        logoIv.setAnimation(animation);
+        animation.setDuration(5000); // Set animation duration to 5 seconds
+        logoIv.setAnimation(animation); // Apply the animation to the logo
+
+        // Play a sound effect using MediaPlayer
         MediaPlayer mediaPlayer = MediaPlayer.create(context, R.raw.swoosh);
         mediaPlayer.start();
+
+        // After a 5-second delay, move to the HomeActivity (main screen)
         new Handler().postDelayed(new Runnable() {
             @Override
             public void run() {
                 Intent intent = new Intent(MainActivity.this, HomeActivity.class);
                 startActivity(intent);
             }
-        }, 5000);
-
-
-
+        }, 5000); // Delay in milliseconds
     }
 }
